Move everything to v0 module
All the "shared" stuff needs re-writing, just move it all to the `v0`
module so that that module stays functional. We can then go wild with
the new v2 stuff.
